+++ packages/flame/test/events/component_mixins/tap_callbacks_test.dart
@@ -233,6 +233,248 @@ void main() {
       },
     );
   });
+
+  testWidgets(
+    'taps are delivered to a TapCallbacks component',
+    (tester) async {
+      var nTapDown = 0;
+      var nLongTapDown = 0;
+      var nTapCancel = 0;
+      var nTapUp = 0;
+      final game = FlameGame(
+        children: [
+          _TapWithCallbacksComponent(
+            size: Vector2(200, 100),
+            position: Vector2(50, 50),
+            onTapDown: (e) => nTapDown++,
+            onLongTapDown: (e) => nLongTapDown++,
+            onTapCancel: (e) => nTapCancel++,
+            onTapUp: (e) => nTapUp++,
+          ),
+        ],
+      );
+      await tester.pumpWidget(GameWidget(game: game));
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 10));
+      expect(game.children.length, 4);
+
+      // regular tap
+      await tester.tapAt(const Offset(100, 100));
+      await tester.pump(const Duration(milliseconds: 100));
+      expect(nTapDown, 1);
+      expect(nTapUp, 1);
+      expect(nLongTapDown, 0);
+      expect(nTapCancel, 0);
+
+      // long tap
+      await tester.longPressAt(const Offset(100, 100));
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDown, 2);
+      expect(nTapUp, 2);
+      expect(nLongTapDown, 1);
+      expect(nTapCancel, 0);
+
+      // cancelled tap
+      var gesture = await tester.startGesture(const Offset(100, 100));
+      await gesture.cancel();
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDown, 3);
+      expect(nTapUp, 2);
+      expect(nTapCancel, 1);
+
+      // tap cancelled via movement
+      gesture = await tester.startGesture(const Offset(100, 100));
+      await gesture.moveBy(const Offset(20, 20));
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDown, 4);
+      expect(nTapUp, 2);
+      expect(nTapCancel, 2);
+    },
+  );
+
+  testWidgets(
+    'TapCallbacks component nested in another TapCallbacks component',
+    (tester) async {
+      var nTapDownChild = 0;
+      var nTapDownParent = 0;
+      var nTapCancelChild = 0;
+      var nTapCancelParent = 0;
+      var nTapUpChild = 0;
+      var nTapUpParent = 0;
+      final game = FlameGame(
+        children: [
+          _TapWithCallbacksComponent(
+            size: Vector2.all(100),
+            position: Vector2.zero(),
+            onTapDown: (e) => nTapDownParent++,
+            onTapUp: (e) => nTapUpParent++,
+            onTapCancel: (e) => nTapCancelParent++,
+            children: [
+              _TapWithCallbacksComponent(
+                size: Vector2.all(50),
+                position: Vector2.all(25),
+                onTapDown: (e) {
+                  nTapDownChild++;
+                  e.continuePropagation = true;
+                },
+                onTapCancel: (e) => nTapCancelChild++,
+                onTapUp: (e) => nTapUpChild++,
+              ),
+            ],
+          ),
+        ],
+      );
+      await tester.pumpWidget(GameWidget(game: game));
+      await tester.pump();
+      await tester.pump();
+      expect(game.children.length, 4);
+      expect(game.children.elementAt(1).children.length, 1);
+
+      await tester.longPressAt(const Offset(50, 50));
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDownChild, 1);
+      expect(nTapDownParent, 1);
+      expect(nTapUpChild, 1);
+      expect(nTapUpParent, 1);
+      expect(nTapCancelChild, 0);
+      expect(nTapCancelParent, 0);
+
+      // cancelled tap
+      final gesture = await tester.startGesture(const Offset(50, 50));
+      await gesture.cancel();
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDownChild, 2);
+      expect(nTapDownParent, 2);
+      expect(nTapUpChild, 1);
+      expect(nTapUpParent, 1);
+      expect(nTapCancelChild, 1);
+      expect(nTapCancelParent, 1);
+    },
+  );
+
+  testWidgets(
+    'tap events do not propagate down by default',
+    (tester) async {
+      var nTapDownParent = 0;
+      var nTapCancelParent = 0;
+      var nTapUpParent = 0;
+      final game = FlameGame(
+        children: [
+          _TapWithCallbacksComponent(
+            size: Vector2.all(100),
+            position: Vector2.zero(),
+            onTapDown: (e) => nTapDownParent++,
+            onTapUp: (e) => nTapUpParent++,
+            onTapCancel: (e) => nTapCancelParent++,
+            children: [
+              _SimpleTapCallbacksComponent(size: Vector2.all(100)),
+            ],
+          ),
+        ],
+      );
+      await tester.pumpWidget(GameWidget(game: game));
+      await tester.pump();
+      await tester.pump();
+      expect(game.children.length, 4);
+      expect(game.children.elementAt(1).children.length, 1);
+
+      await tester.longPressAt(const Offset(50, 50));
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDownParent, 0);
+      expect(nTapUpParent, 0);
+      expect(nTapCancelParent, 0);
+
+      // cancelled tap
+      final gesture = await tester.startGesture(const Offset(50, 50));
+      await gesture.cancel();
+      await tester.pump(const Duration(seconds: 1));
+      expect(nTapDownParent, 0);
+      expect(nTapUpParent, 0);
+      expect(nTapCancelParent, 0);
+    },
+  );
+
+  testWidgets(
+    'local coordinates during tap events',
+    (tester) async {
+      TapDownEvent? tapDownEvent;
+      final game = FlameGame(
+        children: [
+          PositionComponent(
+            size: Vector2.all(400),
+            position: Vector2.all(10),
+            children: [
+              PositionComponent(
+                size: Vector2(300, 200),
+                scale: Vector2(1.5, 2),
+                position: Vector2.all(40),
+                children: [
+                  _TapWithCallbacksComponent(
+                    size: Vector2(100, 50),
+                    position: Vector2(50, 50),
+                    onTapDown: (e) => tapDownEvent = e,
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ],
+      );
+      await tester.pumpWidget(GameWidget(game: game));
+      await tester.pump();
+      await tester.pump();
+      expect(game.children.length, 4);
+      expect(game.children.elementAt(1).children.length, 1);
+
+      await tester.tapAt(const Offset(200, 200));
+      await tester.pump(const Duration(seconds: 1));
+      expect(tapDownEvent, isNotNull);
+      expect(tapDownEvent!.devicePosition, Vector2(200, 200));
+      expect(tapDownEvent!.canvasPosition, Vector2(200, 200));
+      expect(tapDownEvent!.localPosition, Vector2(50, 25));
+      final trace = tapDownEvent!.renderingTrace.reversed.toList();
+      expect(trace[0], Vector2(50, 25));
+      expect(trace[1], Vector2(100, 75));
+      expect(trace[2], Vector2(190, 190));
+      expect(trace[3], Vector2(200, 200));
+    },
+  );
+}
+
+class _TapWithCallbacksComponent extends PositionComponent with TapCallbacks {
+  _TapWithCallbacksComponent({
+    required Vector2 super.position,
+    required Vector2 super.size,
+    super.children,
+    void Function(TapDownEvent)? onTapDown,
+    void Function(TapDownEvent)? onLongTapDown,
+    void Function(TapUpEvent)? onTapUp,
+    void Function(TapCancelEvent)? onTapCancel,
+  })  : _onTapDown = onTapDown,
+        _onLongTapDown = onLongTapDown,
+        _onTapUp = onTapUp,
+        _onTapCancel = onTapCancel;
+
+  final void Function(TapDownEvent)? _onTapDown;
+  final void Function(TapDownEvent)? _onLongTapDown;
+  final void Function(TapUpEvent)? _onTapUp;
+  final void Function(TapCancelEvent)? _onTapCancel;
+
+  @override
+  void onTapDown(TapDownEvent event) => _onTapDown?.call(event);
+
+  @override
+  void onLongTapDown(TapDownEvent event) => _onLongTapDown?.call(event);
+
+  @override
+  void onTapUp(TapUpEvent event) => _onTapUp?.call(event);
+
+  @override
+  void onTapCancel(TapCancelEvent event) => _onTapCancel?.call(event);
+}
+
+class _SimpleTapCallbacksComponent extends PositionComponent with TapCallbacks {
+  _SimpleTapCallbacksComponent({super.size});
 }
 
 mixin _TapCounter on TapCallbacks {
